+++ src/set.ts
@@ -7,6 +7,10 @@ const isObject = (value: unknown) => (typeof value === "object" || typeof value 
  * "on the way". If non-objects are found, they are replaced with new plain objects. If primitives are used as source
  * they are ignored and returned value is empty object with updated value at given path.
  *
+ * This is still too dynamic in nature to get full TypeScript support. Properties are not typed, return type is unknown.
+ * If you are okay with NOT mutating the object but get new one instead it is recommended to use `immutable-assign`
+ * package instead.
+ *
  * @param {Object} source - source object to mutate
  * @param {string|Array<string>} path - path where value should be stored, written as dot-separated property names or
  * array with property names. Use Array when your keys includes dots.